Move cornerRadius modifier to a separate file
