+++ Sources/ShortcutUI/Extensions/View/View+CornerRadius.swift
@@ -1,0 +1,30 @@
+//
+//  View+CornerRadius.swift
+//  ShortcutUI
+//
+//  Created by Sheikh Bayazid on 2022-12-15.
+//
+
+import SwiftUI
+
+#if !os(macOS)
+private struct RoundedCorner: Shape {
+    var radius: CGFloat = .infinity
+    var corners: UIRectCorner = .allCorners
+
+    func path(in rect: CGRect) -> Path {
+        let path = UIBezierPath(roundedRect: rect,
+                                byRoundingCorners: corners,
+                                cornerRadii: CGSize(width: radius, height: radius))
+        return Path(path.cgPath)
+    }
+}
+#endif
+
+public extension View {
+#if !os(macOS)
+    func cornerRadius(_ radius: CGFloat, corners: UIRectCorner) -> some View {
+        clipShape(RoundedCorner(radius: radius, corners: corners))
+    }
+#endif
+}
