+++ Sources/ShortcutUI/Extensions/View/View+Extension.swift
@@ -67,24 +67,4 @@ public extension View {
     func eraseToAnyView() -> AnyView {
         AnyView(self)
     }
-
-#if !os(macOS)
-    func cornerRadius(_ radius: CGFloat, corners: UIRectCorner) -> some View {
-        clipShape(RoundedCorner(radius: radius, corners: corners))
-    }
-#endif
 }
-
-#if !os(macOS)
-private struct RoundedCorner: Shape {
-    var radius: CGFloat = .infinity
-    var corners: UIRectCorner = .allCorners
-
-    func path(in rect: CGRect) -> Path {
-        let path = UIBezierPath(roundedRect: rect,
-                                byRoundingCorners: corners,
-                                cornerRadii: CGSize(width: radius, height: radius))
-        return Path(path.cgPath)
-    }
-}
-#endif
